// UaRestGateway/uarestgateway.client/src/opcua/models/TimestampsToReturn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPC UA REST API
 * This API provides simple REST based access to an OPC UA server.
 *
 * The version of the OpenAPI document: 1.05.4
 * Contact: office@opcfoundation.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TimestampsToReturn = {
    Source: 0,
    Server: 1,
    Both: 2,
    Neither: 3,
    Invalid: 4
} as const;
export type TimestampsToReturn = typeof TimestampsToReturn[keyof typeof TimestampsToReturn];


export function TimestampsToReturnFromJSON(json: any): TimestampsToReturn {
    return TimestampsToReturnFromJSONTyped(json, false);
}

export function TimestampsToReturnFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimestampsToReturn {
    return json as TimestampsToReturn;
}

export function TimestampsToReturnToJSON(value?: TimestampsToReturn | null): any {
    return value as any;
}